scripts/requirements.txt
sparql-client
